pyproject.toml
[project]
name = "config"
version = "0.1.0"
description = ""
authors = [
    {name = "DoanCongPho",email = "congpho115@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "cachetools (>=6.2.1,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
